Reduce Sass files while still having more than one
One of the hardest parts about setting up Sass build system is getting
it to process both Sass and SCSS files. This is one of the first things
I look for when a build system promises to handle Sass. As such, we
should still have multiple files of both types.